+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [4.1.0] - 2026-03-16
+
 ### Added
 
 - support for max length in `TelephoneNumberInput`.
@@ -711,7 +713,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Created package :tada:
 
-[Unreleased]: https://github.com/neolution-ch/react-hook-form-components/compare/4.0.0...HEAD
+[Unreleased]: https://github.com/neolution-ch/react-hook-form-components/compare/4.1.0...HEAD
 [0.1.2]: https://github.com/neolution-ch/react-hook-form-components/compare/0.1.1...0.1.2
 [0.1.1]: https://github.com/neolution-ch/react-hook-form-components/compare/0.1.0...0.1.1
 [0.1.0]: https://github.com/neolution-ch/react-hook-form-components/releases/tag/0.1.0
@@ -736,6 +738,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 [0.4.0]: https://github.com/neolution-ch/react-hook-form-components/compare/0.3.0...0.4.0
 [0.3.0]: https://github.com/neolution-ch/react-hook-form-components/compare/0.2.0...0.3.0
 [0.2.0]: https://github.com/neolution-ch/react-hook-form-components/releases/tag/0.2.0
+[4.1.0]: https://github.com/neolution-ch/react-hook-form-components/compare/4.0.0...4.1.0
 [4.0.0]: https://github.com/neolution-ch/react-hook-form-components/compare/3.17.0...4.0.0
 [3.17.0]: https://github.com/neolution-ch/react-hook-form-components/compare/3.16.0...3.17.0
 [3.16.0]: https://github.com/neolution-ch/react-hook-form-components/compare/3.15.1...3.16.0